+++ src/mailman/runners/command.py
@@ -53,7 +53,7 @@ class CommandFinder:
         self.command_lines = []
         self.ignored_lines = []
         self.processed_lines = []
-        self.send_response = True
+        results.send_response = True
         # Depending on where the message was destined to, add some implicit
         # commands.  For example, if this was sent to the -join or -leave
         # addresses, it's the same as if 'join' or 'leave' commands were sent
@@ -62,12 +62,12 @@ class CommandFinder:
         subaddress = msgdata.get('subaddress')
         if subaddress == 'join':
             self.command_lines.append('join')
-            self.send_response = False
+            results.send_response = False
             is_address_command = True
         elif subaddress == 'leave':
             self.command_lines.append('leave')
             is_address_command = True
-            self.send_response = False
+            results.send_response = False
         elif subaddress == 'confirm':
             # match with re.DOTALL in case header is folded.
             mo = re.match(config.mta.verp_confirm_regexp,
@@ -75,7 +75,7 @@ class CommandFinder:
             if mo:
                 self.command_lines.append('confirm ' + mo.group('cookie'))
                 is_address_command = True
-                self.send_response = False
+                results.send_response = False
         # Stop processing if the address already contained a valid command
         if is_address_command:
             return
@@ -140,6 +140,7 @@ class Results:
     def __init__(self, charset='us-ascii'):
         self._output = StringIO()
         self.charset = charset
+        self.send_response = True
         print(_("""\
 The results of your email command are provided below.
 """), file=self._output)
@@ -215,7 +216,7 @@ class CommandRunner(Runner):
                 if status == ContinueProcessing.no:
                     break
         # All done. If we don't need to send response, return.
-        if not finder.send_response:
+        if not results.send_response:
             return
         # Strip blank lines and send the response.
         lines = [line.strip() for line in finder.command_lines if line]